# archsmartadm.indiginfoundation.com/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArchSmart API Documentation",
        "description": "Complete API documentation for ArchSmart Real Estate Platform. \r\n * \r\n * **Authentication:**\r\n * 1. First, login using the `/login` endpoint with your credentials\r\n * 2. Copy the token from the response\r\n * 3. Click the 'Authorize' button (🔓) at the top right\r\n * 4. Paste your token in the value field (just the token, without 'Bearer')\r\n * 5. Click 'Authorize' and then 'Close'\r\n * 6. Now you can test all protected endpoints\r\n * \r\n * **Note:** All API requests should use `Accept: application/json` header.",
        "contact": {
            "name": "ArchSmart Support",
            "email": "admin@archsmart.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://archsmart.indiginfoundation.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://archsmartadm.indiginfoundation.com/api",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "/api",
            "description": "Current Domain API Server"
        }
    ],
    "paths": {
        "/admin/agents": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all agents (Admin only)",
                "description": "Retrieve list of all agents with their status and approval information",
                "operationId": "149279b52dbf5076696c1c31641a6039",
                "responses": {
                    "200": {
                        "description": "List of agents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "is_approved": {
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/agents/pending": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get pending agents (Admin only)",
                "description": "Retrieve list of agents awaiting approval",
                "operationId": "2593670a78583ee4bdf607d2dad46496",
                "responses": {
                    "200": {
                        "description": "List of pending agents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "is_approved": {
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/agents/{id}/approve": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve an agent (Admin only)",
                "description": "Approve a pending agent registration",
                "operationId": "274a2d401dbeca5e9e9df11e8952916c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agent User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agent approved successfully"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User is not an agent"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/agents/{id}/suspend": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Suspend an agent (Admin only)",
                "description": "Suspend or deactivate an agent account",
                "operationId": "11ab0369379b69d14794c04610b30027",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agent User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent suspended successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agent suspended successfully"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User is not an agent"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "required": [
                    "id",
                    "name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "User's full name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "description": "User role",
                        "type": "string",
                        "enum": [
                            "admin",
                            "agent",
                            "user"
                        ],
                        "example": "user"
                    },
                    "is_agent": {
                        "description": "Whether user is an agent",
                        "type": "boolean",
                        "example": false
                    },
                    "is_admin": {
                        "description": "Whether user is an admin",
                        "type": "boolean",
                        "example": false
                    },
                    "is_approved": {
                        "description": "Whether user account is approved",
                        "type": "boolean",
                        "example": true
                    },
                    "phone": {
                        "description": "User's phone number",
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "bio": {
                        "description": "User's biography",
                        "type": "string",
                        "example": "Experienced real estate agent with 10+ years",
                        "nullable": true
                    },
                    "company": {
                        "description": "User's company name",
                        "type": "string",
                        "example": "ArchSmart Realty",
                        "nullable": true
                    },
                    "profile_picture": {
                        "description": "Path to profile picture",
                        "type": "string",
                        "example": "profile_pictures/user1.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Account creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token (the token will be automatically prefixed with 'Bearer ')",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}